# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BDC API Documentation",
    "version": "1.0.7",
    "description": "BDC (Business Data Constructor) API 文档",
    "contact": {
      "name": "BDC Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3300",
      "description": "开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/database-connections/{id}/test": {
      "post": {
        "tags": [
          "Database Connections"
        ],
        "operationId": "postDatabaseConnectionsIdTest",
        "summary": "测试数据库连接",
        "description": "测试指定ID的数据库连接是否可用",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "数据库连接ID"
          }
        ],
        "responses": {
          "200": {
            "description": "测试成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "连接测试成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DatabaseConnectionTestResult"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "数据库连接不存在"
          },
          "500": {
            "description": "服务器错误或连接测试失败"
          }
        }
      }
    },
...
    "/api/enums": {
      "post": {
        "summary": "创建枚举",
        "tags": [
          "Enum Management"
        ],
        "operationId": "postEnums",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Enum"
              },
              "example": {
                "code": "system:user:status",
                "name": "用户状态",
                "description": "系统用户状态枚举",
                "options": [
                  {
                    "value": "active",
                    "label": "正常"
                  },
                  {
                    "value": "locked",
                    "label": "锁定"
                  },
                  {
                    "value": "disabled",
                    "label": "禁用"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enum"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "409": {
            "description": "枚举代码已存在"
          }
        }
      },
      "get": {
        "summary": "获取枚举列表",
        "tags": [
          "Enum Management"
        ],
        "operationId": "getEnums",
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "是否只返回启用的枚举"
          },
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "description": "按枚举代码模糊搜索"
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": "按枚举名称模糊搜索"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Enum"
                  }
                },
                "example": [
                  {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "code": "system:user:status",
                    "name": "用户状态"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/enums/{id}": {
      "get": {
        "summary": "根据ID获取枚举",
        "tags": [
          "Enum Management"
        ],
        "operationId": "getEnumsId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "枚举ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enum"
                }
              }
            }
          },
          "404": {
            "description": "枚举不存在"
          }
        }
      },
      "put": {
        "summary": "更新枚举",
        "tags": [
          "Enum Management"
        ],
        "operationId": "putEnumsId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "枚举ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Enum"
              },
              "example": {
                "name": "用户状态",
                "description": "系统用户状态枚举-已更新",
                "options": [
                  {
                    "value": "active",
                    "label": "正常"
                  },
                  {
                    "value": "locked",
                    "label": "锁定"
                  },
                  {
                    "value": "disabled",
                    "label": "已禁用"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enum"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "枚举不存在"
          }
        }
      },
      "delete": {
        "summary": "删除枚举",
        "tags": [
          "Enum Management"
        ],
        "operationId": "deleteEnumsId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "枚举ID"
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "404": {
            "description": "枚举不存在"
          },
          "409": {
            "description": "枚举正在被使用，无法删除"
          }
        }
      }
    },
    "/api/enums/code/{code}": {
      "get": {
        "summary": "根据代码获取枚举",
        "tags": [
          "Enum Management"
        ],
        "operationId": "getEnumsCodeCode",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "枚举代码",
            "example": "system:user:status"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Enum"
                }
              }
            }
          },
          "404": {
            "description": "枚举不存在"
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "Enum": {
        "type": "object",
        "required": [
          "code",
          "name",
          "options"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "枚举ID"
          },
          "code": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_:]*$",
            "description": "枚举代码（使用:分隔的多级结构，如 system:user:status）"
          },
          "name": {
            "type": "string",
            "description": "枚举名称"
          },
          "description": {
            "type": "string",
            "maxLength": 100,
            "description": "枚举描述"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "value",
                "label"
              ],
              "properties": {
                "value": {
                  "type": "string",
                  "description": "枚举值（存储值）"
                },
                "label": {
                  "type": "string",
                  "description": "显示标签"
                },
                "description": {
                  "type": "string",
                  "description": "选项描述"
                },
                "order": {
                  "type": "integer",
                  "description": "排序号（可选）"
                }
              }
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
      "DatabaseConnectionTestResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "测试是否成功"
          },
          "message": {
            "type": "string",
            "description": "测试结果消息"
          },
          "connectionInfo": {
            "type": "object",
            "description": "连接信息",
            "properties": {
              "version": {
                "type": "string",
                "description": "数据库版本"
              },
              "serverTime": {
                "type": "string",
                "format": "date-time",
                "description": "服务器时间"
              },
              "maxConnections": {
                "type": "integer",
                "description": "最大连接数"
              },
              "currentConnections": {
                "type": "integer",
                "description": "当前连接数"
              }
            }
          }
        }
      },
...
    }
  }
}